moved playable to its own package
